backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
PyJWT==2.8.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
PyMySQL==1.1.0